// src/types/generated/ComparaisonArticle.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PrixFournisseur } from "./PrixFournisseur";

export type ComparaisonArticle = { article_id: number, designation: string, code_barre: string | null, fournisseurs: Array<PrixFournisseur>, };